Revert back to int32 types, add validations

+++ api/v1/container_types.go
@@ -69,10 +69,14 @@ const (
 // +k8s:openapi-gen=true
 type ContainerPort struct {
 	// Optional: If specified, this must be a valid port number, 0 < x < 65536.
-	HostPort uint16 `json:"hostPort,omitempty"`
+	// +kubebuilder:validation:Minimum=0
+	// +kubebuilder:validation:Maximum=65535
+	HostPort int32 `json:"hostPort,omitempty"`
 
 	// Required: This must be a valid port number, 0 < x < 65536.
-	ContainerPort uint16 `json:"containerPort"`
+	// +kubebuilder:validation:Minimum=0
+	// +kubebuilder:validation:Maximum=65535
+	ContainerPort int32 `json:"containerPort"`
 
 	// The port to be used, defaults to TCP
 	Protocol PortProtocol `json:"protocol,omitempty"`
