+++ api/v1/endpoint_types.go
@@ -31,7 +31,9 @@ type EndpointSpec struct {
 	Address string `json:"address"`
 
 	// The desired port for the endpoint to run on
-	Port uint16 `json:"port"`
+	// +kubebuilder:validation:Minimum=0
+	// +kubebuilder:validation:Maximum=65535
+	Port int32 `json:"port"`
 }
 
 // EndpointStatus describes the status of a Endpoint
